Created area.py

+++ area.py
@@ -1,5 +1,5 @@
 from math import pi
 rad = input("Enter the radius of the circle:")
-rad1 = int(rad)
+rad1 = float(rad)
 ar = pi*((rad1)**2)
 print("The area of the circle is:",ar)
